fix race for positive control counters usage (#13811)
