+++ ydb/library/accessor/positive_integer.h
@@ -7,13 +7,10 @@ namespace NKikimr {
 
 class TPositiveControlInteger {
 private:
-    ui64 Value = 0;
+    TAtomicCounter Value = 0;
 public:
     TPositiveControlInteger() = default;
-    TPositiveControlInteger(const ui64 value)
-        : Value(value) {
-
-    }
+    TPositiveControlInteger(const ui64 value);
     TPositiveControlInteger(const ui32 value)
         : Value(value) {
 
@@ -30,10 +27,10 @@ public:
     ui64 GetDec() const;
     ui64 Val() const;
     bool operator!() const {
-        return !Value;
+        return !Value.Val();
     }
     operator ui64() const {
-        return Value;
+        return Value.Val();
     }
     ui64 operator++() {
         return Inc();
